ability to change roles

+++ convex/files.ts
@@ -33,8 +33,9 @@ async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, orgId: string) {
     return null;
   }
 
-  const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
+  const hasAccess =user.orgIds.some((item) => item.orgId === orgId) ||
+  user.tokenIdentifier.includes(orgId);
+   
     if (!hasAccess) {
       return null;
     }
@@ -114,7 +115,13 @@ export const deleteFile = mutation({
     if (!access) {
       throw new ConvexError("no access to file");
     }
+    const isAdmin =
+      access.user.orgIds.find((org) => org.orgId === access.file.orgId)
+        ?.role === "admin";
 
+    if (!isAdmin) {
+      throw new ConvexError("you have no admin access to delete");
+    }
     await ctx.db.delete(args.fileId);
   },
 });
